pyproject.toml
[tool.poetry]
name = "tgbot"
version = "0.1.0"
description = ""
authors = ["Ayrelia <nikitakoteqka12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.10.0"
aiohttp = "^3.9.5"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.31"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"
asyncio = "^3.4.3"
apscheduler = "^3.10.4"
cachetools = "^5.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
